requirements.txt
lxml 
requests 
pyaml